Refactor evaluate: Extract generate & reward functions
